refactor: extract access key identity helpers
