+++ rustfs/src/admin/mod.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+pub(crate) mod access_key_identity;
 mod auth;
 pub mod console;
 pub mod handlers;
